+++ Libplanet/Action/ActionContext.cs
@@ -15,10 +15,12 @@ namespace Libplanet.Action
     {
         public static readonly AsyncLocal<Stopwatch> GetStateTimer = new AsyncLocal<Stopwatch>();
         public static readonly AsyncLocal<int> GetStateCount = new AsyncLocal<int>();
+        public static readonly AsyncLocal<GasMeter> GetGasMeter = new AsyncLocal<GasMeter>();
 
         private readonly int _randomSeed;
         private readonly ITrie? _previousBlockStatesTrie;
         private readonly Predicate<Currency>? _nativeTokenPredicate;
+        private readonly long _gasLimit;
         private HashDigest<SHA256>? _previousStateRootHash;
 
         public ActionContext(
@@ -29,6 +31,7 @@ namespace Libplanet.Action
             long blockIndex,
             IAccountStateDelta previousStates,
             int randomSeed,
+            long gasLimit,
             bool rehearsal = false,
             ITrie? previousBlockStatesTrie = null,
             bool blockAction = false,
@@ -44,6 +47,7 @@ namespace Libplanet.Action
             PreviousStates = previousStates;
             Random = new Random(randomSeed);
             _randomSeed = randomSeed;
+            _gasLimit = gasLimit;
             _previousBlockStatesTrie = previousBlockStatesTrie;
             BlockAction = blockAction;
             _nativeTokenPredicate = nativeTokenPredicate;
@@ -51,6 +55,7 @@ namespace Libplanet.Action
 
             GetStateTimer.Value = new Stopwatch();
             GetStateCount.Value = 0;
+            GetGasMeter.Value = new GasMeter(_gasLimit);
         }
 
         public BlockHash? GenesisHash { get; }
@@ -100,10 +105,41 @@ namespace Libplanet.Action
                 BlockIndex,
                 PreviousStates,
                 _randomSeed,
+                _gasLimit,
                 Rehearsal,
                 _previousBlockStatesTrie,
                 BlockAction,
                 _nativeTokenPredicate,
                 new List<string>());
+
+        /// <summary>
+        /// Returns the elapsed gas of the current action.
+        /// </summary>
+        /// <returns>
+        /// The elapsed gas of the current action.
+        /// </returns>
+        /// <exception cref="InvalidOperationException">
+        /// Thrown when <see cref="GetGasMeter"/> is not initialized.
+        /// </exception>
+        public long GasUsed()
+        {
+            if (GetGasMeter.Value is { } gasMeter)
+            {
+                return gasMeter.GasUsed;
+            }
+
+            throw new InvalidOperationException($"{nameof(GetGasMeter)} is not initialized.");
+        }
+
+        /// <summary>
+        /// Returns the gas limit of the current action.
+        /// </summary>
+        /// <returns>
+        /// The gas limit of the current action.
+        /// </returns>
+        public long GasLimit()
+        {
+            return _gasLimit;
+        }
     }
 }